Refactoring DB files
Reorganized the structure of the database files to avoid circular dependencies being formed.
