+++ sciencelabs/course/__init__.py
@@ -4,7 +4,7 @@ from flask_classy import FlaskView
 
 # Local
 from sciencelabs.course.course_controller import CourseController
-from sciencelabs.db_repository.CourseRepo import Course
+from sciencelabs.db_repository.course_functions import CourseFunctions
 
 
 class CourseView(FlaskView):
@@ -14,5 +14,5 @@ class CourseView(FlaskView):
         self.base = CourseController()
 
     def index(self):
-        course_info = Course().get_course_info()
+        course_info = CourseFunctions().get_course_info()
         return render_template('course/base.html', **locals())
