--- sciencelabs/db_repository/CourseProfessorsRepo.py
@@ -1,10 +1,0 @@
-from sqlalchemy import *
-
-from sciencelabs.db_repository import Base
-from sciencelabs.db_repository import session
-
-
-class CourseProfessors(Base):
-    __tablename__ = 'CourseProfessors'
-    course_id = Column(Integer, primary_key=True)
-    professor_id = Column(Integer)
