--- sciencelabs/db_repository/CourseViewerRepo.py
@@ -1,10 +1,0 @@
-from sqlalchemy import *
-
-from sciencelabs.db_repository import Base
-from sciencelabs.db_repository import session
-
-
-class CourseViewer(Base):
-    __tablename__ = 'CourseViewer'
-    course_id = Column(Integer, primary_key=True)
-    user_id = Column(Integer)
