+++ sciencelabs/schedule/__init__.py
@@ -5,7 +5,7 @@ from datetime import datetime
 
 # Local
 from sciencelabs.schedule.schedule_controller import ScheduleController
-from sciencelabs.db_repository.ScheduleRepo import Schedule
+from sciencelabs.db_repository.schedule_functions import ScheduleFunctions
 
 
 class ScheduleView(FlaskView):
@@ -14,7 +14,7 @@ class ScheduleView(FlaskView):
 
     def index(self):
         timedelta_to_time = datetime.min
-        schedule_info = Schedule().get_schedule_tab_info()
+        schedule_info = ScheduleFunctions().get_schedule_tab_info()
         return render_template('schedule/base.html', **locals())
 
     @route('/create')
