+++ sciencelabs/session/__init__.py
@@ -5,9 +5,8 @@ from datetime import datetime
 
 # Local
 from sciencelabs.session.session_controller import SessionController
-from sciencelabs.db_repository.SessionRepo import Session
-from sciencelabs.db_repository.TutorSessionRepo import TutorSession
-from sciencelabs.db_repository.UserRepo import User
+from sciencelabs.db_repository.session_functions import SessionFunctions
+from sciencelabs.db_repository.user_functions import UserFunctions
 
 
 class SessionView(FlaskView):
@@ -19,8 +18,8 @@ class SessionView(FlaskView):
 
     def closed(self):
         timedelta_to_time = datetime.min
-        sessions = Session().get_closed_sessions()
-        session_tutors = TutorSession()
+        sessions = SessionFunctions().get_closed_sessions()
+        session_tutors = SessionFunctions()
         return render_template('session/closed_sessions.html', **locals())
 
     def create(self):
@@ -31,9 +30,9 @@ class SessionView(FlaskView):
 
     def edit_session(self, session_id):
         timedelta_to_time = datetime.min
-        session = Session.get_session(self, session_id)
-        leads, tutors = TutorSession.get_session_tutors(self, session_id)
-        session_students = User.get_session_students(self, session_id)
+        session = SessionFunctions().get_session(session_id)
+        leads, tutors = SessionFunctions().get_session_tutors(session_id)
+        session_students = UserFunctions().get_session_students(session_id)
         return render_template('session/edit_closed_session.html', **locals())
 
     def edit_student(self):
@@ -47,7 +46,7 @@ class SessionView(FlaskView):
 
     def edit_tutor(self, tutor_id, session_id):
         timedelta_to_time = datetime.min
-        tutor = TutorSession.get_tutor_session_info(self, tutor_id, session_id)
+        tutor = SessionFunctions().get_tutor_session_info(tutor_id, session_id)
         return render_template('session/edit_tutor.html', **locals())
 
     def add_tutor(self):
